+++ src/main/resources/static/js/map_geo.js
@@ -6,118 +6,9 @@ var city = $("meta[name='defined_city']").attr("content");
 var latitude;
 var longitude;
 
-codeAddress();
-
-function codeAddress() {
-    geocoder = new google.maps.Geocoder;
-    geocoder.geocode({'address': city}, function (results, status) {
-        if (status == google.maps.GeocoderStatus.OK) {
-            latitude = results[0].geometry.location.lat();
-            longitude = results[0].geometry.location.lng();
-            initMap();
-        }
-    });
-}
-
-
 function initMap() {
-    if(latitude===undefined){
-        codeAddress();
-    }
-    let styledMapType = new google.maps.StyledMapType(
-        [
-            {
-                "elementType": "geometry",
-                "stylers": [{"color": "#f5f5f5"}]
-            },
-            {
-                "elementType": "labels.icon",
-                "stylers": [{"visibility": "off"}]
-            },
-            {
-                "elementType": "labels.text.fill",
-                "stylers": [{"color": "#616161"}]
-            },
-            {
-                "elementType": "labels.text.stroke",
-                "stylers": [{"color": "#f5f5f5"}]
-            },
-            {
-                "featureType": "administrative.land_parcel", "elementType": "labels.text.fill",
-                "stylers": [{"color": "#bdbdbd"}]
-            },
-            {
-                "featureType": "poi",
-                "elementType": "geometry",
-                "stylers": [{"color": "#eeeeee"}]
-            },
-            {
-                "featureType": "poi",
-                "elementType": "labels.text.fill",
-                "stylers": [{"color": "#757575"}]
-            },
-            {
-                "featureType": "poi.park",
-                "elementType": "geometry",
-                "stylers": [{"color": "#0adc51"}]
-            },
-            {
-                "featureType": "poi.park",
-                "elementType": "labels.text.fill",
-                "stylers": [{"color": "#9e9e9e"}]
-            },
-            {
-                "featureType": "road",
-                "elementType": "geometry",
-                "stylers": [{"color": "#ffc107"}]
-            },
-            {
-                "featureType": "road.arterial",
-                "elementType": "labels.text.fill",
-                "stylers": [{"color": "#757575"}]
-            },
-            {
-                "featureType": "road.highway",
-                "elementType": "geometry",
-                "stylers": [{"color": "#ff9907"}]
-            },
-            {
-                "featureType": "road.highway",
-                "elementType": "labels.text.fill",
-                "stylers": [{"color": "#616161"}]
-            },
-            {
-                "featureType": "road.local",
-                "elementType": "labels.text.fill",
-                "stylers": [{"color": "#9e9e9e"}]
-            },
-            {
-                "featureType": "transit.line",
-                "elementType": "geometry",
-                "stylers": [{"color": "#e5e5e5"}]
-            },
-            {
-                "featureType": "transit.station",
-                "elementType": "geometry",
-                "stylers": [{"color": "#eeeeee"}]
-            },
-            {
-                "featureType": "water",
-                "elementType": "geometry",
-                "stylers": [{"color": "#4d90fe"}]
-                // /* water color styled map: 007bff*/
-            },
-            {
-                "featureType": "water",
-                "elementType": "labels.text.fill",
-                "stylers": [{"color": "#9e9e9e"}]
-            }
-        ], {name: 'Styled Map'});
-
-    var viborg = {lat: latitude, lng: longitude};
-
     map = new google.maps.Map(document.getElementById('map'), {
-        center: viborg,
+        // center: viborg,
         zoom: 14,
         gestureHandling: 'cooperative',
         streetViewControl: false,
@@ -129,7 +20,11 @@ function initMap() {
         }
     });
 
+    geocoder = new google.maps.Geocoder;
+    codeAddress();
+
     //Associate the styled map with the MapTypeId and set it to display.
+    let styledMapType = new google.maps.StyledMapType(styledMap, {name: 'Styled Map'});
     map.mapTypes.set('styled_map', styledMapType);
     map.setMapTypeId('styled_map');
 
@@ -193,6 +88,18 @@ function initMap() {
     let centerControl = new CenterControl(centerControlDiv, map);
     centerControlDiv.index = 1;
     map.controls[google.maps.ControlPosition.RIGHT_CENTER].push(centerControlDiv);
+}
+
+function codeAddress() {
+    let viborg = {lat: 60.704958391204, lng: 28.753046876075};
+
+    geocoder.geocode({'address': city}, function (results, status) {
+        if (status == google.maps.GeocoderStatus.OK) {
+            map.setCenter(results[0].geometry.location);
+        } else {
+            map.setCenter(viborg);
+        }
+    });
 }
 
 function createInfoOfferTable(tableID) {
