+++ src/main/resources/static/js/styled_map.js
@@ -1,4 +1,90 @@
-var styledMapType;
-function initMap() {
 
-}
+var styledMap =
+    [
+        {
+            "elementType": "geometry",
+            "stylers": [{"color": "#f5f5f5"}]
+        },
+        {
+            "elementType": "labels.icon",
+            "stylers": [{"visibility": "off"}]
+        },
+        {
+            "elementType": "labels.text.fill",
+            "stylers": [{"color": "#616161"}]
+        },
+        {
+            "elementType": "labels.text.stroke",
+            "stylers": [{"color": "#f5f5f5"}]
+        },
+        {
+            "featureType": "administrative.land_parcel", "elementType": "labels.text.fill",
+            "stylers": [{"color": "#bdbdbd"}]
+        },
+        {
+            "featureType": "poi",
+            "elementType": "geometry",
+            "stylers": [{"color": "#eeeeee"}]
+        },
+        {
+            "featureType": "poi",
+            "elementType": "labels.text.fill",
+            "stylers": [{"color": "#757575"}]
+        },
+        {
+            "featureType": "poi.park",
+            "elementType": "geometry",
+            "stylers": [{"color": "#0adc51"}]
+        },
+        {
+            "featureType": "poi.park",
+            "elementType": "labels.text.fill",
+            "stylers": [{"color": "#9e9e9e"}]
+        },
+        {
+            "featureType": "road",
+            "elementType": "geometry",
+            "stylers": [{"color": "#ffc107"}]
+        },
+        {
+            "featureType": "road.arterial",
+            "elementType": "labels.text.fill",
+            "stylers": [{"color": "#757575"}]
+        },
+        {
+            "featureType": "road.highway",
+            "elementType": "geometry",
+            "stylers": [{"color": "#ff9907"}]
+        },
+        {
+            "featureType": "road.highway",
+            "elementType": "labels.text.fill",
+            "stylers": [{"color": "#616161"}]
+        },
+        {
+            "featureType": "road.local",
+            "elementType": "labels.text.fill",
+            "stylers": [{"color": "#9e9e9e"}]
+        },
+        {
+            "featureType": "transit.line",
+            "elementType": "geometry",
+            "stylers": [{"color": "#e5e5e5"}]
+        },
+        {
+            "featureType": "transit.station",
+            "elementType": "geometry",
+            "stylers": [{"color": "#eeeeee"}]
+        },
+        {
+            "featureType": "water",
+            "elementType": "geometry",
+            "stylers": [{"color": "#4d90fe"}]
+            // /* water color styled map: 007bff*/
+        },
+        {
+            "featureType": "water",
+            "elementType": "labels.text.fill",
+            "stylers": [{"color": "#9e9e9e"}]
+        }
+    ];
