+++ src/main/resources/templates/new_offer.html
@@ -7,11 +7,13 @@
     <title>MainPage</title>
     <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyDhmsuY2GsI23ECoyUqAaVtWVDz3GleyTo&libraries=places&callback=initMap"
             async defer></script>
+    <meta name="city" th:content="${city}"/>
     <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous"></script>
     <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script>
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
     <script type="text/javascript" src="/js/sender/new_offer.js"></script>
+    <script type="text/javascript" src="/js/styled_map.js"></script>
     <!--<script type="text/javascript" src="/js/sender/point_location.js"></script>-->
     <style></style>
 </head>
@@ -22,6 +24,7 @@
         <div class="row">
             <div class="col-sm-4"></div>
             <div class="col-sm-4" align="center" style="color: #7295B1">
+                <h2> Добавление новой точки вывоза</h2>
                 <form action="/sender/new_offer" method="post">
                     <div class="card-body">
                         <div class="form-group">
@@ -95,19 +98,12 @@
                         <label class="col control-label">Указать адрес вывоза</label>
                         <div class="col-md-12 inputGroupContainer">
                             <div class="input-group">
-                                <div class="input-group-prepend">
-                                        <span class="input-group-btn">
-                                            <button class="btn btn-primary" type="button" data-toggle="modal" data-target="#mapModal"
-                                                    id="modalMap" style="background-color: #F39C12; color: #222" onclick="initV_Map()">
-                                                <i class="fas fa-map"></i>
-                                            </button>
-                                        </span>
-                                </div>
-                                <input name="address" id="address" placeholder="<- открыть карту"
-                                       class="form-control" type="text" readonly>
-                                <input type="hidden" value="123" name="latitude"/>
-                                <input type="hidden" value="123" name="longitude"/>
+                                <input name="address" id="address" placeholder="Введите полный адрес вручную или выберите на карте"
+                                       class="form-control" type="text">
+                                <input id="o_latitude" type="hidden" value="123" name="latitude"/>
+                                <input id="o_longitude" type="hidden" value="123" name="longitude"/>
                             </div>
+                            <div id="map" style="height: 600px"></div>
                         </div>
                     </div>
 
@@ -123,7 +119,7 @@
                         </div>
 
                         <div class="form-group" align="center">
-                            <button type="submit" class="btn btn-warning" style="width: 250px;">
+                            <button type="submit" id="send_offer" class="btn btn-warning" style="width: 250px;">
                                 Зарегистрироваться <i class="fas fa-paper-plane "></i>
                             </button>
                         </div>
@@ -131,26 +127,6 @@
                     </div>
                 </form>
             </div>
-        </div>
-    </div>
-
-    <!--Modal-->
-    <div id="mapModal" class="modal fade" role="dialog">
-        <div class="modal-dialog">
-            <!-- Modal content-->
-            <div class="modal-content">
-                <div class="modal-header">
-                    <h2 class="modal-title">Укажите место на карте</h2>
-                    <button type="button" class="close" data-dismiss="modal">&times;</button>
-                </div>
-                <div class="modal-body">
-                    <input name="address_modal" id="address_modal"
-                           placeholder="Введите полный адрес вручную или выберите на карте"
-                           class="form-control" type="text">
-                </div>
-                <div id="map" style="height: 600px"></div>
-            </div>
-
         </div>
     </div>
 </div>
